backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "learning phase"
authors = [{ name = "Zahin_Abdullah", email = "zahinabdullahrakin@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "beanie (>=1.29.0,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "bcrypt (<4.1.0)",
    "google-auth (>=2.39.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "datetime (>=5.5,<6.0)",
]

[tool.poetry.scripts]
dev = "backend.main:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
